Added support for ExpressionEngine
Uses google "im feeling lucky" feature to search the EE user guide for
a given keyword, and go directly to the top result. Not perfect but it
will get you to the right page most of the time, and improving this
would likely require EE user guide to be reorganized around predictable
URLs for each keyword.